[solidago] init package configuration, move mehestan primitives
