+++ backend/dev-env/Dockerfile
@@ -4,9 +4,3 @@ ENV PYTHONUNBUFFERED=1
 WORKDIR /backend
 
 RUN apt-get update && apt-get install -y gettext && rm -rf /var/lib/apt/lists/*
-
-COPY requirements.txt /backend/
-COPY tests/requirements.txt /backend/requirements.dev.txt
-
-RUN pip install -r requirements.txt
-RUN pip install -r requirements.dev.txt
