--- backend/tests/test_addition.py
@@ -1,6 +1,0 @@
-import pytest
-
-
-@pytest.mark.unit
-def test_addition():
-    assert (1 + 1) == 2
