+++ dev-env/docker-compose.yml
@@ -31,12 +31,14 @@ services:
     entrypoint: ["/bin/bash", "-c"]
     command:
       - |
+        pip install -r requirements.dev.txt -r requirements.txt
         python manage.py migrate
         python manage.py createcachetable
         django-admin compilemessages
         python manage.py runserver 0.0.0.0:8000
     volumes:
       - ../backend:/backend
+      - ../solidago:/solidago
     ports:
       - "127.0.0.1:8000:8000"
     networks:
