+++ solidago/README.md
@@ -1,0 +1,3 @@
+# Solidago
+
+Algorithms for the Tournesol platform
