+++ solidago/src/solidago/__init__.py
@@ -1,0 +1,3 @@
+"""Solidago library, robust and secure algorithms for the Tournesol platform"""
+
+__version__ = "0.0.1"
