+++ solidago/src/solidago/optimize.py
